+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Added `terminal?` function
+* `Decider.compose(left, right)`
 
 # 0.1.0